Move codegen util files out of typePolicies/
